requirements.txt
eel
scipy==1.15.3
antspyx
nibabel
dipy==1.9.0
numpy==1.26.4
pyinstaller
matplotlib
scikit-image
pyside6
click
customtkinter
pillow
